requirements.txt
openai
pillow
opencv-python
moviepy
fastapi
uvicorn
gTTS
englisttohindi
pyttsx3
python-dotenv
transformers
torch